Split out SignatureHandler into its own mod
